Python Strings

+++ 0x07-python-test_driven_development/102-python.c
@@ -11,7 +11,7 @@
 
 void print_python_string(PyObject *p)
 {
-    Py_ssize_t length = ((PyASCIIObject *)(p))->length;
+    long int length = ((PyASCIIObject *)(p))->length;
     const wchar_t *wide_str = PyUnicode_AsWideCharString(p, &length);
 
 	fflush(stdout);
@@ -25,7 +25,7 @@ void print_python_string(PyObject *p)
     printf("[.] string object info\n");
 
     printf("  type: %s\n", PyUnicode_IS_COMPACT_ASCII(p) ? "compact ascii" : "compact unicode object");
-    printf("  length: %zd\n", length);
+    printf("  length: %ld\n", length);
     printf("  value: %ls\n", wide_str);
 
     PyMem_Free((void *)wide_str);
